+++ CHANGELOG.md
@@ -48,8 +48,6 @@ All notable changes to this project will be documented in this file.
 - Fix npm publish
 - Test build
 - Removed test
-- *(tauri-plugin-webauthn)* Release v0.1.0
+- _(tauri-plugin-webauthn)_ Release v0.1.0
 - Added deps for release plz
 - Removed changelog
-
-